refactor: move useFormatNumberByNucleus hook
