+++ src/component/2d/tools/XYLabelPointer.tsx
@@ -3,7 +3,7 @@ import { useContext, useMemo, CSSProperties } from 'react';
 import { BrushContext } from '../../EventsTrackers/BrushTracker';
 import { MouseContext } from '../../EventsTrackers/MouseTracker';
 import { useChartData } from '../../context/ChartContext';
-import { useFormatNumberByNucleus } from '../../utility/formatNumber';
+import { useFormatNumberByNucleus } from '../../hooks/useFormatNumberByNucleus';
 import { getLayoutID, LAYOUT } from '../utilities/DimensionLayout';
 import { get2DXScale, get2DYScale, get1DYScale } from '../utilities/scale';
